requirements.txt
Django>=4.2
djangorestframework>=3.14
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.26.0,<0.27
djangorestframework-simplejwt==5.5.0
django-cors-headers>=4.3.1
Pillow>=8.2.0,<8.3.0
django-filter>=25.1
redis>=4.5.1,<4.6.0
hiredis>=2.2.0,<2.3.0
django-redis>=5.2.0,<5.3.0
